Create README.md
Trolololol trololol trollin 4 u! Trollface is a black and white drawing of a face with a large mischievous grin that is meant to portray the expression someone makes while trolling. Posting a Trollface image into a forum thread is often used to claim that someone was being fooled or intentionally angered. The face commonly appears in rage comic indicating that the character is being mischievous in some way. The original drawing was created by deviantArt[ user Whynne on September 19th, 2008 to illustrate the pointless nature of trolling on 4chan’s /v/ (videogames) board. Trollface (also known as "Cool Face") is a meme associated with 4chan and other so-called "troll" sites - normally it is little more than a mascot for trolls but there are many comics that show Trollface to be a living creature (often with a stick-figure body) and he is symbolic of internet trolls and the chan culture. TroUMAD trolllface is often shown to do incredibly stupid, evil or obnoxious things and delights in mischief - he is sometimes portrayed as having a rivalry of sorts with "Epic Fail Guy" or in other "stories" they are allies. Like most internet villains Trollface does not usually have much in the way of logical plot of storyline - though some of his exploits can be considered a running series as people make webcomics and movies based around the character. Series such as "Troll Science" featured Trollface quite prominently and he can frequently be seen in many humor-sites such as "Funny Junk", the Chans and (formerly) Encyclopedia Dramatica: his appearance can vary creator. n Internet slang, a troll (/ˈtroʊl/, /ˈtrɒl/) is a person who sows discord on the Internet by starting arguments or upsetting people, by posting inflammatory, extraneous, or off-topic messages in an online community (such as a forum, chat room, or blog), either accidentally or with the deliberate intent of provoking readers into an emotional response or of otherwise disrupting normal on-topic discussion. The first post Trollface was used in. The original Trollface post Trollface is thought by many to come from sites such as 4chan, YTMND, Something Awful, ED, 2ch or any other site like these but in fact it comes from a very talented user on DevianArt called Whynne